Move app-logic to hooks folder
